src/AVTextify.Python/requirements.txt
torch
git+https://github.com/openai/whisper.git
ffmpeg-python
pandas